src/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.4.0
aiofile==3.8.8 #for using file chunk by chunk
